fastapi-ip-scanner/app/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
python-dotenv
pydantic
